#196 - 카메라 UI 작업 완료
