+++ WorkoutDone/WorkoutDone/Resources/Assets.xcassets/Home/Camera/Contents.json
@@ -1,0 +1,6 @@
+{
+  "info" : {
+    "author" : "xcode",
+    "version" : 1
+  }
+}
