+++ WorkoutDone/WorkoutDone/Sources/Presenter/HomeScene/Camera/HomeButtonLess/DeniedCameraHomeButtonLessView.swift
@@ -1,0 +1,8 @@
+//
+//  DeniedCameraHomeButtonLessView.swift
+//  WorkoutDone
+//
+//  Created by 류창휘 on 2023/05/23.
+//
+
+import Foundation
